Production/requirements.txt
pdf2image == 1.17.0
pytesseract == 0.3.10
pillow == 10.3.0